backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
SQLAlchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.1
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4

# Google Cloud
google-cloud-bigquery==3.12.0
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.36.4
google-cloud-logging==3.8.0
google-auth==2.23.4
google-auth-oauthlib==0.8.0
google-auth-httplib2==0.1.1

# Data processing
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
# TensorFlow for ML (optional - can use BigQuery ML instead)
# tensorflow==2.14.0

# API integrations
requests==2.31.0
aiohttp==3.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Development
black==23.10.1
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0